requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
dlib>=19.24.2
face-recognition>=1.3.0
scikit-learn>=1.3.0
Pillow>=10.0.0
Click>=8.0.0
Flask>=3.0.0
faiss-cpu>=1.7.4
